1.0.2
完善了用户查询，添加了分页和token时效验证等功能

+++ spring/src/main/java/org/example/springbootdemo/dto/UserDTO.java
@@ -7,7 +7,7 @@ public class UserDTO {
     private Integer id;
     private String name;
     private Integer money=0;
-    private Integer status;
-    private Integer role_id;
+    private Byte status;
+    private Byte role_id;
     private String password;
 }
