+++ spring/src/main/java/org/example/springbootdemo/vo/UserVO.java
@@ -8,6 +8,6 @@ public class UserVO {
     private Integer id;
     private String name;
     private Integer money;
-    private Integer status;
-    private Integer role_id;
+    private Byte status;
+    private Byte role_id;
 }
